# frontend/backoffice/src/lib/apigen/openapi.yaml
openapi: 3.0.3
info:
  title: CATL Tour API
  description: |
    REST API for the CATL logistics route-preparation system.
    Manages producers, their vehicles, customers, and delivery routes (tours)
    with their ordered stops. Provides helper services for geocoding, routing,
    route optimization and KPI reporting.
  version: 0.1.0
  contact:
    name: CATL
servers:
  - url: http://localhost:8080/api/v1
    description: Local dev
  - url: https://api.catl.example/api/v1
    description: Production
...
paths:
...
  /infrastructures:
    get:
      tags: [Infrastructure]
      summary: List infrastructures
      operationId: listInfrastructures
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Page of infrastructures
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InfrastructurePage' }
    post:
      tags: [Infrastructure]
      summary: Create an infrastructure
      operationId: createInfrastructure
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/InfrastructureCreate' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Infrastructure' }
  /infrastructures/{infrastructureId}:
    parameters:
      - $ref: '#/components/parameters/InfrastructureId'
    get:
      tags: [Infrastructure]
      summary: Get an infrastructure
      operationId: getInfrastructure
      responses:
        '200':
          description: Infrastructure
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Infrastructure' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [Infrastructure]
      summary: Update an infrastructure
      operationId: updateInfrastructure
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/InfrastructureUpdate' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Infrastructure' }
    delete:
      tags: [Infrastructure]
      summary: Delete an infrastructure
      operationId: deleteInfrastructure
      responses:
        '204': { description: Deleted }

  # ── Vehicles ──────────────────────────────────────────────
...
components:
...
  parameters:
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    Offset:
      name: offset
      in: query
      schema: { type: integer, minimum: 0, default: 0 }
...
    InfrastructureId:
      name: infrastructureId
      in: path
      required: true
      schema: { type: string, format: uuid }

  # ── Reusable responses ────────────────────────────────────
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:    { type: string, example: VALIDATION_FAILED }
        message: { type: string, example: "Field 'name' must not be blank" }
        details:
          type: object
          additionalProperties: true
    Page:
      type: object
      required: [total, limit, offset]
      properties:
        total:  { type: integer, example: 42 }
        limit:  { type: integer, example: 50 }
        offset: { type: integer, example: 0 }

    # ── Enums ─────────────────────────────────────────────
...
    Infrastructure:
      type: object
      required: [id, drySurfaceM2, freshSurfaceM2, frozenSurfaceM2, prepSurfaceM2]
      properties:
        id:              { type: string, format: uuid, readOnly: true }
        drySurfaceM2:    { type: integer, minimum: 0 }
        freshSurfaceM2:  { type: integer, minimum: 0 }
        frozenSurfaceM2: { type: integer, minimum: 0 }
        prepSurfaceM2:   { type: integer, minimum: 0 }
        depotLatitude:   { type: number, format: double, minimum: -90,  maximum: 90,  nullable: true }
        depotLongitude:  { type: number, format: double, minimum: -180, maximum: 180, nullable: true }
        updatedAt:       { type: string, format: date-time, readOnly: true }
    InfrastructureCreate:
      type: object
      required: [drySurfaceM2, freshSurfaceM2, frozenSurfaceM2, prepSurfaceM2]
      properties:
        drySurfaceM2:    { type: integer, minimum: 0, default: 0 }
        freshSurfaceM2:  { type: integer, minimum: 0, default: 0 }
        frozenSurfaceM2: { type: integer, minimum: 0, default: 0 }
        prepSurfaceM2:   { type: integer, minimum: 0, default: 0 }
        depotLatitude:   { type: number, format: double, minimum: -90,  maximum: 90,  nullable: true }
        depotLongitude:  { type: number, format: double, minimum: -180, maximum: 180, nullable: true }
    InfrastructureUpdate:
      type: object
      properties:
        drySurfaceM2:    { type: integer, minimum: 0 }
        freshSurfaceM2:  { type: integer, minimum: 0 }
        frozenSurfaceM2: { type: integer, minimum: 0 }
        prepSurfaceM2:   { type: integer, minimum: 0 }
        depotLatitude:   { type: number, format: double, minimum: -90,  maximum: 90,  nullable: true }
        depotLongitude:  { type: number, format: double, minimum: -180, maximum: 180, nullable: true }
    InfrastructurePage:
      allOf:
        - $ref: '#/components/schemas/Page'
        - type: object
          properties:
            items:
              type: array
              items: { $ref: '#/components/schemas/Infrastructure' }

    # ── Vehicle ───────────────────────────────────────────
...